backend/requirements.txt
fastapi==0.109.2
uvicorn==0.30.3
psycopg2-binary==2.9.9
python-multipart==0.0.9
